lib/requirements.txt
pandas 
numpy 
matplotlib 
seaborn 
python-docx 
openpyxl